refactor: code split and lazy load new ui and old ui
